pyproject.toml
[project]
name = "uni"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "bcrypt (==4.0.1)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "openai>=2.6.1,<3.0.0",
    "openai-agents (>=0.4.2,<0.5.0)",
    "rich (>=14.2.0,<15.0.0)",
    "chainlit (>=2.9.2,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "fastmcp (>=2.14.0,<3.0.0)",
]

[tool.poetry.scripts]
start = "uni.main:start"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
